ecommerce-scraper/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic[email]==2.5.3
python-dotenv==1.0.0
email-validator==2.1.0
python-multipart==0.0.6
cryptography==41.0.7

# Chart generation libraries
matplotlib==3.8.2
seaborn==0.13.1
pandas==2.1.4
numpy==1.26.3
plotly 
kaleido
apscheduler